project-inventory/src/main/java/com/enigma/arinda/entities/StockSummary.java
package com.enigma.arinda.entities;

public class StockSummary {

    private Item item;
    private Long quantity;
    private Long totalPrice;

    public StockSummary() {
    }

    public StockSummary(Item item, Long quantity, Long totalPrice) {
        this.item = item;
        this.quantity = quantity;
        this.totalPrice = totalPrice;
    }

    public Item getItem() {
        return item;
    }

    public void setItem(Item item) {
        this.item = item;
    }

    public Long getQuantity() {
        return quantity;
    }

    public void setQuantity(Long quantity) {
        this.quantity = quantity;
    }

    public Long getTotalPrice() {
        return totalPrice;
    }

    public void setTotalPrice(Long totalPrice) {
        this.totalPrice = totalPrice;
    }

}

